+++ src/main/java/io/github/lambdaphoenix/simpleterminal/ansi/AnsiColor.java
@@ -16,7 +16,7 @@ package io.github.lambdaphoenix.simpleterminal.ansi;
  * }</pre>
  *
  * @author lambdaphoenix
- * @version 2025-09-25
+ * @version 0.2.0 (2025-10-18)
  * @since 0.1.0
  */
 public final class AnsiColor {