+++ src/main/java/io/github/lambdaphoenix/simpleterminal/ansi/package-info.java
@@ -22,7 +22,7 @@
  * }</pre>
  *
  * @author lambdaphoenix
- * @version 2025-09-25
+ * @version 0.2.0 (2025-10-18)
  * @since 0.1.0
  */
 package io.github.lambdaphoenix.simpleterminal.ansi;